requirements.txt
discord.py-self>=1.9.2
PyNaCl>=1.5.0
